+++ src/components/config/constants.ts
@@ -1,0 +1,4 @@
+export const HOME_URL = '/home';
+export const CHAT_URL = '/chat';
+export const GAME_URL = '/pong';
+export const SETTING_URL = '/settings';
